// models/code-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bandwidth
 * Bandwidth\'s Communication APIs
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: letstalk@bandwidth.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CodeRequest
 */
export interface CodeRequest {
    /**
     * The phone number to send the mfa code to.
     * @type {string}
     * @memberof CodeRequest
     */
    'to': string;
    /**
     * The application phone number, the sender of the mfa code.
     * @type {string}
     * @memberof CodeRequest
     */
    'from': string;
    /**
     * The application unique ID, obtained from Bandwidth.
     * @type {string}
     * @memberof CodeRequest
     */
    'applicationId': string;
    /**
     * An optional field to denote what scope or action the mfa code is addressing.  If not supplied, defaults to \"2FA\".
     * @type {string}
     * @memberof CodeRequest
     */
    'scope'?: string;
    /**
     * The message format of the mfa code.  There are three values that the system will replace \"{CODE}\", \"{NAME}\", \"{SCOPE}\".  The \"{SCOPE}\" and \"{NAME} value template are optional, while \"{CODE}\" must be supplied.  As the name would suggest, code will be replace with the actual mfa code.  Name is replaced with the application name, configured during provisioning of mfa.  The scope value is the same value sent during the call and partitioned by the server.
     * @type {string}
     * @memberof CodeRequest
     */
    'message': string;
    /**
     * The number of digits for your mfa code.  The valid number ranges from 2 to 8, inclusively.
     * @type {number}
     * @memberof CodeRequest
     */
    'digits': number;
}